--- src/component/Categories.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Categories = () => (
-  <div>
-    <button className="btn-check-status"type="button">Check Status</button>
-  </div>
-);
-
-export default Categories;
