+++ src/components/Book.js
@@ -1,4 +1,4 @@
-import PropTypes from "prop-types";
+import PropTypes from 'prop-types';
 
 const Book = (props) => {
   const { title, author } = props;
